+++ src/app/app-users.js
@@ -1,7 +1,8 @@
 const fastifyPlugin = require('fastify-plugin');
 
+const loginRoute = require('./users/login');
 const getAllUsersRoute = require('./users/get-users');
-const createUserRoute = require('./auth/register');
+const createUserRoute = require('./users/register');
 const getUserRoute = require('./users/get-user');
 const getUserRolesRoute = require('./users/get-user-roles');
 
@@ -10,7 +11,8 @@ async function plugin(fastify, opts, next) {
 		.register(getAllUsersRoute, opts)
 		.register(createUserRoute, opts)
 		.register(getUserRoute, opts)
-		.register(getUserRolesRoute, opts);
+		.register(getUserRolesRoute, opts)
+		.register(loginRoute, opts);
 
 	fastify.setNotFoundHandler((req, res) => {
 		const {username} = req.params;
